pyproject.toml
[tool.poetry]
name = "sigstickers"
version = "2024.1"
license = "mit"
description = "Download sticker packs from Signal"
authors = ["FredHappyface"]
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"Intended Audience :: Education",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: Implementation :: CPython",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Utilities",
	"Topic :: Multimedia :: Graphics",
]
homepage = "https://github.com/FHPythonUtils/SigStickers"
repository = "https://github.com/FHPythonUtils/SigStickers"
documentation = "https://github.com/FHPythonUtils/SigStickers/blob/master/README.md"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "<11,>=10.2.0"
signalstickers-client = "<4,>=3.3.0"
emoji = "<3,>=2.10.1"
loguru = "<2,>=0.7.2"

[tool.poetry.scripts]
sigstickers = "sigstickers:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pylint = "^3.1.0"
handsdown = "^2.1.0"
coverage = "^7.4.4"
pytest-asyncio = "^0.23.5.post1"

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
	"ANN101",  # type annotation for self in method
	"COM812",  # enforce trailing comma
	"D2",  # pydocstyle formatting
	"ISC001",
	"N",  # pep8 naming
	"PLR09",  # pylint refactor too many
	"TCH",  # type check blocks
	"W191"  # ignore this to allow tabs
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["D", "S101", "E402"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all" # Disallow all relative imports.

[tool.ruff.format]
indent-style = "tab"
docstring-code-format = true
line-ending = "lf"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.tox]
legacy_tox_ini = """
[tox]
env_list =
	py311
	py310
	py39
	py38

[testenv]
deps = pytest
commands = pytest tests
"""
